Refactored abstract database service to simpler extension database ty…

+++ abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/multivisualisation/service/impl/ChartConfigurationServiceImpl.java
@@ -106,7 +106,8 @@ public class ChartConfigurationServiceImpl implements ChartConfigurationService 
     public void removeChartConfiguration(Long moduleId) {
         log.debug("removeChartConfiguration - moduleId", moduleId);
         ChartConfiguration chartConfiguration = chartConfigurationRepository.findByModuleId(moduleId);
-        chartConfigurationRepository.delete(chartConfiguration);
-
+        if (chartConfiguration != null) {
+            chartConfigurationRepository.delete(chartConfiguration);
+        }
     }
 }
